backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
mysqlclient==2.2.5
asgiref==3.8.1
sqlparse==0.5.1
openpyxl==3.1.5
xhtml2pdf==0.2.16

